requirements.txt
ipykernel==6.29.5
python-dotenv==1.0.1
requests==2.32.3
langchain==0.3.17
langchain-chroma==0.2.1
langchain-community==0.3.16
sentence-transformers==3.4.1
langchain-google-genai==2.0.9
fastapi==0.115.8
pydantic==2.10.6
SQLAlchemy==2.0.38